refactor: separate color select component
